requirements.txt
streamlit
chromadb
pypdf
google-genai
python-dotenv